+++ smos-docs-site/content/pages/changelog.markdown
@@ -9,6 +9,7 @@ description: The changelog for all of the Smos tools and libraries
 
 - `smos-github`: There is now a new column that suggests you to have another look at the issue/pr because there has been a recent enough update.
 - `smos-github`: Done entries are no longer mentioned in the `smos-github list` report.
+- `smos`: A keybinding to open the `url` property in smos: `<space>ou` by default.
 
 # <a name="2021-04-09">[2021-04-09 Release](#2021-04-09)